+++ src/components/pages/Fatality/FatalityCard.js
